codex - reuse docs drawer
