+++ docs/src/pages/Installation.tsx
@@ -2,14 +2,15 @@
 // src/pages/Installation.tsx  | valet
 // Getting started installation page
 // ─────────────────────────────────────────────────────────────
-import { Surface, Stack, Typography, Button, Panel } from '@archway/valet';
+import { Stack, Typography, Button, Panel } from '@archway/valet';
+import DocsLayout from './DocsLayout';
 import { useNavigate } from 'react-router-dom';
 
 export default function InstallationPage() {
   const navigate = useNavigate();
 
   return (
-    <Surface>
+    <DocsLayout>
       <Stack spacing={1} preset="showcaseStack">
         <Typography variant="h2" bold>Installation</Typography>
         <Typography>Install via npm:</Typography>
@@ -19,6 +20,6 @@ export default function InstallationPage() {
         <Typography>For now, valet works best with React 18</Typography>
         <Button onClick={() => navigate(-1)}>← Back</Button>
       </Stack>
-    </Surface>
+    </DocsLayout>
   );
 }
